Functions now create and populate symbol tables
However I need to look into how to add the function identifier to the
appropriate symbol table...as it is now, while processing semantic
records I pop the identifier off the list, but we need it to both insert
as a symbol and to create a new table.  I will look into this.